@@ -1,7 +1,9 @@
-module.exports = {
-  semi: false,
-  singleQuote: true,
-  trailingComma: 'all',
+/** @type {import('prettier').Config} */
+export default {
   printWidth: 100,
+  singleQuote: false,
+  semi: true,
+  trailingComma: "all",
   tabWidth: 2,
-}
+  arrowParens: "always"
+};
